cartitems fetch now inside cart summary
